fix (DPLAN-12665) Create statementPhaseService
